rename: CourseListDrawer 를 component/map 에서 component/post 폴더로 이동
